fix(cmd): preserve fractional request timeouts
Convert timeout seconds after scaling to nanoseconds so values like 0.5s are not truncated to zero.\n\nThis keeps serve mode and normal CLI initialization consistent for sub-second request timeouts.

+++ cmd/helpers.go
@@ -99,7 +99,7 @@ func initClient(ctx context.Context) (*client.Client, map[string]string, error) 
 		AccountIndex:  acctIdx,
 		Model:         model,
 		Verbose:       verbose,
-		Timeout:       time.Duration(requestTimeout) * time.Second,
+		Timeout:       time.Duration(requestTimeout * float64(time.Second)),
 	}
 
 	c, err := client.New(cfg)
